Task_2/requirements.txt
torch==2.2.0 --index-url https://download.pytorch.org/whl/cu124
torchvision==0.17.0 --index-url https://download.pytorch.org/whl/cu124
transformers>=4.30.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
Pillow>=9.4.0
datasets>=2.18.0
numpy>=1.24.0
tqdm>=4.65.0